+++ src/components/dummies/b-dummy/b-dummy.ts
@@ -11,10 +11,11 @@
  * @packageDocumentation
  */
 
+import { components } from 'core/component/const';
 import type { VNode } from 'core/component/engines';
 
 import type iBlock from 'components/super/i-block/i-block';
-import iData, { component, field } from 'components/super/i-data/i-data';
+import iData, { component, field, computed } from 'components/super/i-data/i-data';
 
 export * from 'components/super/i-data/i-data';
 
@@ -24,28 +25,79 @@ export * from 'components/super/i-data/i-data';
 	}
 })
 
-class bDummy extends iData {
+export default class bDummy extends iData {
 	/**
-	 * Name of the test component
+	 * The name of the test component
 	 */
 	@field()
 	testComponent?: string;
 
 	/**
-	 * Attributes for the test component
+	 * Attributes of the test component
 	 */
 	@field()
 	testComponentAttrs: Dictionary = {};
 
 	/**
-	 * Slots for the test component
+	 * Slots of the test component
 	 */
 	@field()
 	testComponentSlots?: CanArray<VNode>;
 
+	/**
+	 * A proxy to modify property values that will be passed to a child component
+	 */
+	@field()
+	storedProps: Dictionary = {};
+
+	/**
+	 * Cached accessors that will be provided as props to the child component
+	 */
+	@field()
+	storedAccessors: Dictionary = {};
+
 	protected override readonly $refs!: iData['$refs'] & {
 		testComponent?: iBlock;
 	};
-}
 
-export default bDummy;
+	/**
+	 * Returns normalized attributes for the test component
+	 */
+	@computed({dependencies: ['testComponentAttrs']})
+	protected get testComponentAttrsNormalized(): Dictionary {
+		const meta = components.get(this.testComponent ?? '');
+
+		if (meta == null) {
+			return this.testComponentAttrs;
+		}
+
+		// Creating new object to prevent mutation of the field
+		const attrs = {};
+
+		Object.keys(this.testComponentAttrs).forEach((key) => {
+			const value = this.testComponentAttrs[key];
+
+			if (meta.props[key]?.forceUpdate === false) {
+				if (key in this.storedProps) {
+					this.storedProps[key] = value;
+					attrs[`@:${key}`] = this.storedAccessors[key];
+
+				} else {
+					this.storedProps[key] = value;
+
+					const
+						accessor = this.createPropAccessors(() => <object>this.storedProps[key]),
+						normalizedKey = `@:${key}`;
+
+					this.storedAccessors[key] = accessor;
+					attrs[normalizedKey] = accessor;
+				}
+
+			} else {
+				attrs[key] = value;
+			}
+		});
+
+		return attrs;
+	}
+}