+++ src/components/super/i-block/mods/index.ts
@@ -27,18 +27,14 @@ export default abstract class iBlockMods extends iBlockEvent {
 	override readonly mods!: ModsDict;
 
 	@computed({dependencies: ['mods']})
-	override get sharedMods(): CanNull<Readonly<ModsDict>> {
-		const
-			m = this.mods;
-
-		let
-			res: CanUndef<ModsDict>;
+	override get sharedMods(): CanNull<ModsDict> {
+		const m = this.mods;
 
 		if (m.theme != null) {
-			res = {theme: m.theme};
+			return {theme: m.theme};
 		}
 
-		return res != null ? Object.freeze(res) : null;
+		return null;
 	}
 
 	/**