+++ src/core/component/decorators/prop/test/b-effect-prop-wrapper-dummy/index.js
@@ -1,0 +1,11 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+package('b-effect-prop-wrapper-dummy')
+	.extends('b-dummy')
+	.dependencies('b-effect-prop-dummy', 'b-non-effect-prop-dummy');